requirements.txt
asgiref==3.8.1
Django==5.0.4
django-crispy-forms==2.1
django-jazzmin==2.6.1
django-widget-tweaks==1.5.0
mysqlclient==2.2.4
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
celery==5.3.6
djangorestframework==3.15.1
